rf-prediction-service/requirements.txt
pandas
numpy
scikit-learn
joblib
google-cloud-bigquery
google-cloud-storage
tzdata
db-dtypes